Expose phased dispatcher start & stop
Today, the dispatcher exposes its lifecycle via all-in-one start and
stop methods. This requires that users start and stop all transports,
inbounds, and outbounds at the same time. In our internal Fx module,
this makes it impossible to correctly order `OnStart` and `OnStop`
hooks: ideally, the dispatcher would start all transports and outbounds,
then consumers of the `ClientConfig` would start (including all
user-supplied procedures and their dependencies), and then the
dispatcher would start inbounds. This requires registering hooks for the
transports and outbounds during dispatcher construction and registering
hooks for inbounds during procedure registration.

This commit exposes two new methods on the dispatcher: `PhasedStart` and
`PhasedStop`. Each method returns a builder-style struct that strongly
suggests the correct startup and shutdown order. PhasedStart and Start
are mutually exclusive, as are PhasedStop and Stop. Unlike the Start and
Stop methods, the builders aren't safe for concurrent or repeated use
(though I'm happy to discuss changing that if it's controversial).